refactor(namespaces): restructure console namespaces
Updated namespaces in ConsoleGameRunner.cs and SystemConsoleIO.cs to better organize game-related components and input/output operations. The Program.cs file was also modified to reflect these changes, ensuring a clearer association with game functionalities.

BREAKING CHANGE: The namespace for ConsoleGameRunner and SystemConsoleIO has changed, which may affect their usage in other parts of the codebase.
