+++ MooGameV3/Presentation/Console/Game/ConsoleGameRunner.cs
@@ -1,12 +1,11 @@
 ﻿using MooGameV3.Application.Abstractions;
 using MooGameV3.Application.Services;
 using MooGameV3.Domain.Game;
-using MooGameV3.Presentation.Console.Game;
 using MooGameV3.Presentation.Console.Intro;
 using MooGameV3.Presentation.Console.IO;
 using MooGameV3.Presentation.Console.Scoring;
 
-namespace MooGameV3.Presentation.Console;
+namespace MooGameV3.Presentation.Console.Game;
 
 public sealed class ConsoleGameRunner(
 	IGameIO io,
